exercise 18
Give the dimensions of the next 2 rectangles in the sequence. Write down
a formula that describes the n-th element in this sequence. Write down a
contract, purpose statement, examples, and definition for a function
rec-sequence that takes an argument n, where n is a number that tells
the function to return the nth element in this sequence. Test the
function!