Exercise 21
Ex21: Here is a struct definition:

(define-struct person (first-name last-name age height weight))
Write down a reasonable comment part of that data definition that
specifies types and interpretations of the fields. Then write the
function person-image that takes a person and returns an image like the
ones in Ex20, but in a way that the height and width of this image is
related to the height of the person (i.e. if one person is twice the
size of another person, the image for the first person should be twice
as high and wide as the image of the second person).